generalize OKCoinWebsocket actor
